Add HW1 Reference File
Added the HW1 reference file.  I had to translate some syntax (since he uses Python 2.7 instead of 3.6) but I confirmed the same error rates that he posted on the course webpage so be sure to use this version.  Also, our in-house algorithms do not match his error rates  and it seems like the error is somewhere within the featuresBinarized file.  Though, I'm not willing to search for it, so I'm game with just using his reference file and calling it a day. -Peter